+++ src/Telegram.Bot/Converters/UnixDatetimeConverter.cs
@@ -3,9 +3,7 @@ using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-#if !NET46
 using Telegram.Bot.Helpers;
-#endif
 
 namespace Telegram.Bot.Converters
 {
@@ -28,11 +26,7 @@ namespace Telegram.Bot.Converters
             long val;
             if (value is DateTime || value is Nullable<DateTime>)
             {
-#if NET46
-                val = new DateTimeOffset((DateTime)value).ToUnixTimeSeconds();
-#else
                 val = ((DateTime)value).ToUnixTime();
-#endif
             }
             else
             {
@@ -66,11 +60,7 @@ namespace Telegram.Bot.Converters
 
             var ticks = (long)reader.Value;
 
-#if NET46
-            return DateTimeOffset.FromUnixTimeSeconds(ticks).DateTime;
-#else
             return ticks.FromUnixTime();
-#endif
         }
 
         /// <summary>