+++ src/Telegram.Bot/TelegramBotClient.cs
@@ -13,6 +13,7 @@ using System.Threading.Tasks;
 using Newtonsoft.Json;
 
 using Telegram.Bot.Args;
+using Telegram.Bot.Converters;
 using Telegram.Bot.Exceptions;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
@@ -36,6 +37,19 @@ namespace Telegram.Bot
         private bool _invalidToken;
         private readonly HttpClient _httpClient;
 
+        private static readonly JsonSerializerSettings SerializerSettings = new JsonSerializerSettings
+        {
+            Converters = new List<JsonConverter>
+            {
+                new ChatIdConverter(),
+                new FileToSendConverter(),
+                new InlineQueryResultTypeConverter(),
+                new ParseModeConverter(),
+                new PhotoSizeConverter(),
+                new UnixDateTimeConverter(),
+            },
+        };
+
         #region Config Properties
 
         /// <summary>
@@ -43,12 +57,12 @@ namespace Telegram.Bot
         /// </summary>
         public TimeSpan Timeout
         {
-            get { return _httpClient.Timeout; }
-            set { _httpClient.Timeout = value; }
+            get => _httpClient.Timeout;
+            set => _httpClient.Timeout = value;
         }
 
         /// <summary>
-        /// Indecates if receiving updates
+        /// Indicates if receiving updates
         /// </summary>
         public bool IsReceiving { get; set; }
 
@@ -101,7 +115,7 @@ namespace Telegram.Bot
         public event EventHandler<UpdateEventArgs> OnUpdate;
 
         /// <summary>
-        /// Occurs when a <see cref="Message"/> is recieved.
+        /// Occurs when a <see cref="Message"/> is received.
         /// </summary>
         public event EventHandler<MessageEventArgs> OnMessage;
 
@@ -126,12 +140,12 @@ namespace Telegram.Bot
         public event EventHandler<CallbackQueryEventArgs> OnCallbackQuery;
 
         /// <summary>
-        /// Occurs when an error occures during the background update pooling.
+        /// Occurs when an error occurs during the background update pooling.
         /// </summary>
         public event EventHandler<ReceiveErrorEventArgs> OnReceiveError;
 
         /// <summary>
-        /// Occurs when an error occures during the background update pooling.
+        /// Occurs when an error occurs during the background update pooling.
         /// </summary>
         public event EventHandler<ReceiveGeneralErrorEventArgs> OnReceiveGeneralError;
 
@@ -142,7 +156,7 @@ namespace Telegram.Bot
         /// </summary>
         /// <param name="token">API token</param>
         /// <param name="httpClient">A custom <see cref="HttpClient"/></param>
-        /// <exception cref="ArgumentException">Thrown if <paramref name="token"/> format is invvalid</exception>
+        /// <exception cref="ArgumentException">Thrown if <paramref name="token"/> format is invalid</exception>
         public TelegramBotClient(string token, HttpClient httpClient = null)
         {
             if (!Regex.IsMatch(token, @"^\d*:[\w\d-_]{35}$"))
@@ -157,7 +171,7 @@ namespace Telegram.Bot
         /// </summary>
         /// <param name="token">API token</param>
         /// <param name="webProxy">Use this <see cref="IWebProxy"/> to connect to the API</param>
-        /// <exception cref="ArgumentException">Thrown if <paramref name="token"/> format is invvalid</exception>
+        /// <exception cref="ArgumentException">Thrown if <paramref name="token"/> format is invalid</exception>
         public TelegramBotClient(string token, IWebProxy webProxy)
         {
             if (!Regex.IsMatch(token, @"^\d*:[\w\d-_]{35}$"))
@@ -172,7 +186,6 @@ namespace Telegram.Bot
             _token = token;
             _httpClient = new HttpClient(httpClientHander);
         }
-
 
         #region Helpers
 
@@ -639,6 +652,7 @@ namespace Telegram.Bot
         /// <param name="chatId"><see cref="ChatId"/> for the target chat</param>
         /// <param name="videoNote">Video note to send.</param>
         /// <param name="duration">Duration of sent video in seconds</param>
+        /// <param name="length">Video width and height</param>
         /// <param name="disableNotification">Sends the message silently. iOS users will not receive a notification, Android users will receive a notification with no sound.</param>
         /// <param name="replyToMessageId">If the message is a reply, ID of the original message</param>
         /// <param name="replyMarkup">Additional interface options. A JSON-serialized object for a custom reply keyboard, instructions to hide keyboard or to force a reply from the user.</param>
@@ -647,15 +661,19 @@ namespace Telegram.Bot
         /// <see href="https://core.telegram.org/bots/api#sendvideonote"/>
         public Task<Message> SendVideoNoteAsync(ChatId chatId, FileToSend videoNote,
             int duration = 0,
+            int length = 0,
             bool disableNotification = false,
             int replyToMessageId = 0,
             IReplyMarkup replyMarkup = null,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            var additionalParameters = new Dictionary<string, object>
-            {
-                {"duration", duration},
-            };
+            var additionalParameters = new Dictionary<string, object>();
+
+            if (duration > 0)
+                additionalParameters.Add("duration", duration);
+
+            if (length > 0)
+                additionalParameters.Add("length", length);
 
             return SendMessageAsync(MessageType.VideoNoteMessage, chatId, videoNote, disableNotification, replyToMessageId,
                 replyMarkup, additionalParameters, cancellationToken);
@@ -806,7 +824,7 @@ namespace Telegram.Bot
         /// <param name="fileId">File identifier</param>
         /// <param name="destination">The destination stream</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>The File object. If <paramref name="destination"/> is empty stream ist embedded in the <see cref="File"/> Object</returns>
+        /// <returns>The File object. If <paramref name="destination"/> stream in not provided, stream is embedded in the <see cref="File"/> object</returns>
         /// <see href="https://core.telegram.org/bots/api#getfile"/>
         public async Task<File> GetFileAsync(string fileId, Stream destination = null,
             CancellationToken cancellationToken = default(CancellationToken))
@@ -840,17 +858,22 @@ namespace Telegram.Bot
         /// </summary>
         /// <param name="chatId"><see cref="ChatId"/> for the target group</param>
         /// <param name="userId">Unique identifier of the target user</param>
+        /// <param name="untilDate"><see cref="DateTime"/> when the user will be unbanned. If user is banned for more than 366 days or less than 30 seconds from the current time they are considered to be banned forever</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <returns><c>true</c> on success.</returns>
         /// <see href="https://core.telegram.org/bots/api#kickchatmember"/>
         public Task<bool> KickChatMemberAsync(ChatId chatId, int userId,
+            DateTime untilDate = default(DateTime),
             CancellationToken cancellationToken = default(CancellationToken))
         {
             var parameters = new Dictionary<string, object>
             {
                 {"chat_id", chatId},
-                {"user_id", userId}
+                {"user_id", userId},
             };
+
+            if (untilDate != default(DateTime))
+                parameters.Add("until_date", untilDate);
 
             return SendWebRequestAsync<bool>("kickChatMember", parameters, cancellationToken);
         }
@@ -1007,6 +1030,94 @@ namespace Telegram.Bot
             return SendWebRequestAsync<bool>("answerCallbackQuery", parameters, cancellationToken);
         }
 
+        /// <summary>
+        /// Use this method to restrict a user in a supergroup. The bot must be an administrator in the supergroup for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target supergroup</param>
+        /// <param name="userId">Unique identifier of the target user</param>
+        /// <param name="untilDate"><see cref="DateTime"/> when restrictions will be lifted for the user. If user is restricted for more than 366 days or less than 30 seconds from the current time, they are considered to be restricted forever</param>
+        /// <param name="canSendMessages">Pass True, if the user can send text messages, contacts, locations and venues</param>
+        /// <param name="canSendMediaMessages">Pass True, if the user can send audios, documents, photos, videos, video notes and voice notes, implies can_send_messages</param>
+        /// <param name="canSendOtherMessages">Pass True, if the user can send animations, games, stickers and use inline bots, implies can_send_media_messages</param>
+        /// <param name="canAddWebPagePreviews">Pass True, if the user may add web page previews to their messages, implies can_send_media_messages</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>On success, <c>true</c> is returned</returns>
+        /// <remarks>Pass True for all boolean parameters to lift restrictions from a user.</remarks>
+        /// <see href="https://core.telegram.org/bots/api#restrictchatmember"/>
+        public Task<bool> RestrictChatMemberAsync(ChatId chatId, int userId, DateTime untilDate = default(DateTime),
+            bool? canSendMessages = null, bool? canSendMediaMessages = null, bool? canSendOtherMessages = null,
+            bool? canAddWebPagePreviews = null,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId },
+                { "user_id", userId }
+            };
+
+            if (untilDate != default(DateTime))
+                parameters.Add("until_date", untilDate);
+
+            if (canSendMessages != null)
+                parameters.Add("can_send_messages", canSendMessages.Value);
+            if (canSendMediaMessages != null)
+                parameters.Add("can_send_media_messages", canSendMediaMessages.Value);
+            if (canSendOtherMessages != null)
+                parameters.Add("can_send_other_messages", canSendOtherMessages.Value);
+            if (canAddWebPagePreviews != null)
+                parameters.Add("can_add_web_page_previews", canAddWebPagePreviews.Value);
+
+            return SendWebRequestAsync<bool>("restrictChatMember", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to promote or demote a user in a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target channel</param>
+        /// <param name="userId">Unique identifier of the target user</param>
+        /// <param name="canChangeInfo">Pass True, if the administrator can change chat title, photo and other settings</param>
+        /// <param name="canPostMessages">Pass True, if the administrator can create channel posts, channels only</param>
+        /// <param name="canEditMessages">Pass True, if the administrator can edit messages of other users, channels only</param>
+        /// <param name="canDeleteMessages">Pass True, if the administrator can delete messages of other users</param>
+        /// <param name="canInviteUsers">Pass True, if the administrator can invite new users to the chat</param>
+        /// <param name="canRestrictMembers">Pass True, if the administrator can restrict, ban or unban chat members</param>
+        /// <param name="canPinMessages">Pass True, if the administrator can pin messages, supergroups only</param>
+        /// <param name="canPromoteMembers">Pass True, if the administrator can add new administrators with a subset of his own privileges or demote administrators that he has promoted, directly or indirectly (promoted by administrators that were appointed by him)</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns True on success.</returns>
+        /// <remarks>Pass False for all boolean parameters to demote a user.</remarks>
+        /// <see href="https://core.telegram.org/bots/api#promotechatmember"/>
+        public Task<bool> PromoteChatMemberAsync(ChatId chatId, int userId, bool? canChangeInfo = null,
+            bool? canPostMessages = null, bool? canEditMessages = null, bool? canDeleteMessages = null,
+            bool? canInviteUsers = null, bool? canRestrictMembers = null, bool? canPinMessages = null,
+            bool? canPromoteMembers = null, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId },
+                { "user_id", userId }
+            };
+
+            if (canChangeInfo != null)
+                parameters.Add("can_change_info", canChangeInfo.Value);
+            if (canPostMessages != null)
+                parameters.Add("can_post_messages", canPostMessages.Value);
+            if (canEditMessages != null)
+                parameters.Add("can_edit_messages", canEditMessages.Value);
+            if (canDeleteMessages != null)
+                parameters.Add("can_delete_messages", canDeleteMessages.Value);
+            if (canInviteUsers != null)
+                parameters.Add("can_invite_users", canInviteUsers.Value);
+            if (canRestrictMembers != null)
+                parameters.Add("can_restrict_members", canRestrictMembers.Value);
+            if (canPinMessages != null)
+                parameters.Add("can_pin_messages", canPinMessages.Value);
+            if (canPromoteMembers != null)
+                parameters.Add("can_promote_members", canPromoteMembers.Value);
+
+            return SendWebRequestAsync<bool>("promoteChatMember", parameters, cancellationToken);
+        }
+
         #endregion Available methods
 
         #region Updating messages
@@ -1055,9 +1166,9 @@ namespace Telegram.Bot
         /// <param name="disableWebPagePreview">Disables link previews for links in this message</param>
         /// <param name="replyMarkup">A JSON-serialized object for an inline keyboard.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>On success, the edited Message is returned.</returns>
+        /// <returns>><c>true</c> on success.</returns>
         /// <see href="https://core.telegram.org/bots/api#editmessagetext"/>
-        public Task<Message> EditInlineMessageTextAsync(string inlineMessageId, string text,
+        public Task<bool> EditInlineMessageTextAsync(string inlineMessageId, string text,
             ParseMode parseMode = ParseMode.Default,
             bool disableWebPagePreview = false,
             IReplyMarkup replyMarkup = null,
@@ -1076,7 +1187,7 @@ namespace Telegram.Bot
             if (parseMode != ParseMode.Default)
                 parameters.Add("parse_mode", parseMode.ToModeString());
 
-            return SendWebRequestAsync<Message>("editMessageText", parameters, cancellationToken);
+            return SendWebRequestAsync<bool>("editMessageText", parameters, cancellationToken);
         }
 
         /// <summary>
@@ -1113,9 +1224,9 @@ namespace Telegram.Bot
         /// <param name="caption">New caption of the message</param>
         /// <param name="replyMarkup">A JSON-serialized object for an inline keyboard.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>On success, the edited Message is returned.</returns>
+        /// <returns>><c>true</c> on success.</returns>
         /// <see href="https://core.telegram.org/bots/api#editmessagecaption"/>
-        public Task<Message> EditInlineMessageCaptionAsync(string inlineMessageId, string caption,
+        public Task<bool> EditInlineMessageCaptionAsync(string inlineMessageId, string caption,
             IReplyMarkup replyMarkup = null,
             CancellationToken cancellationToken = default(CancellationToken))
         {
@@ -1128,7 +1239,7 @@ namespace Telegram.Bot
             if (replyMarkup != null)
                 parameters.Add("reply_markup", replyMarkup);
 
-            return SendWebRequestAsync<Message>("editMessageCaption", parameters, cancellationToken);
+            return SendWebRequestAsync<bool>("editMessageCaption", parameters, cancellationToken);
         }
 
         /// <summary>
@@ -1162,9 +1273,9 @@ namespace Telegram.Bot
         /// <param name="inlineMessageId">Unique identifier of the sent message</param>
         /// <param name="replyMarkup">A JSON-serialized object for an inline keyboard.</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>On success, the edited Message is returned.</returns>
+        /// <returns>><c>true</c> on success.</returns>
         /// <see href="https://core.telegram.org/bots/api#editmessagereplymarkup"/>
-        public Task<Message> EditInlineMessageReplyMarkupAsync(string inlineMessageId,
+        public Task<bool> EditInlineMessageReplyMarkupAsync(string inlineMessageId,
             IReplyMarkup replyMarkup = null,
             CancellationToken cancellationToken = default(CancellationToken))
         {
@@ -1176,7 +1287,7 @@ namespace Telegram.Bot
             if (replyMarkup != null)
                 parameters.Add("reply_markup", replyMarkup);
 
-            return SendWebRequestAsync<Message>("editMessageReplyMarkup", parameters, cancellationToken);
+            return SendWebRequestAsync<bool>("editMessageReplyMarkup", parameters, cancellationToken);
         }
 
         /// <summary>
@@ -1446,7 +1557,7 @@ namespace Telegram.Bot
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <returns>On success, if the message was sent by the bot, returns the edited Message</returns>
         /// <see href="https://core.telegram.org/bots/api#setgamescore"/>
-        public Task<Message> SetGameScoreAsync(int userId, int score, string inlineMessageId,
+        public async Task<Message> SetGameScoreAsync(int userId, int score, string inlineMessageId,
             bool force = false,
             bool disableEditMessage = false,
             bool editMessage = false,
@@ -1462,7 +1573,22 @@ namespace Telegram.Bot
                 {"edit_message", editMessage}
             };
 
-            return SendWebRequestAsync<Message>("setGameScore", parameters, cancellationToken);
+            object response = await SendWebRequestAsync<object>("setGameScore", parameters, cancellationToken)
+                .ConfigureAwait(false);
+
+            Message message;
+            switch (response)
+            {
+                case Message m:
+                    message = m;
+                    break;
+                case bool b when b:
+                    message = default(Message);
+                    break;
+                default:
+                    throw new Exception($"Unexpected response: {response}");
+            }
+            return message;
         }
 
         /// <summary>
@@ -1516,6 +1642,284 @@ namespace Telegram.Bot
 
         #endregion Games
 
+        #region Group and channel management
+        /// <summary>
+        /// Use this method to export an invite link to a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target channel</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns exported invite link as String on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#exportchatinvitelink"/>
+        public Task<string> ExportChatInviteLinkAsync(ChatId chatId,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId }
+            };
+
+            return SendWebRequestAsync<string>("exportChatInviteLink", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to set a new profile photo for the chat. Photos can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target channel</param>
+        /// <param name="photo">The new profile picture for the chat.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns <c>true</c> on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#setchatphoto"/>
+        public Task<bool> SetChatPhotoAsync(ChatId chatId, FileToSend photo,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId },
+                { "photo", photo }
+            };
+
+            return SendWebRequestAsync<bool>("setChatPhoto", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to delete a chat photo. Photos can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target channel</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns true on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#deletechatphoto"/>
+        public Task<bool> DeleteChatPhotoAsync(ChatId chatId, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId }
+            };
+
+            return SendWebRequestAsync<bool>("deleteChatPhoto", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to change the title of a chat. Titles can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target channel</param>
+        /// <param name="title">New chat title, 1-255 characters</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns true on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#setchattitle"/>
+        public Task<bool> SetChatTitleAsync(ChatId chatId, string title,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId },
+                { "title", title }
+            };
+
+            return SendWebRequestAsync<bool>("setChatTitle", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to change the description of a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target channel</param>
+        /// <param name="description">New chat description, 0-255 characters. Defaults to an empty string, which would clear the description.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns true on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#setchatdescription"/>
+        public Task<bool> SetChatDescriptionAsync(ChatId chatId, string description = "",
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId }
+            };
+
+            if (!string.IsNullOrEmpty(description))
+                parameters.Add("description", description);
+
+            return SendWebRequestAsync<bool>("setChatDescription", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to pin a message in a supergroup. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target supergroup</param>
+        /// <param name="messageId">Identifier of a message to pin</param>
+        /// <param name="disableNotification">Pass True, if it is not necessary to send a notification to all group members about the new pinned message</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns true on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#pinchatmessage"/>
+        public Task<bool> PinChatMessageAsync(ChatId chatId, int messageId, bool disableNotification = false,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId },
+                { "message_id", messageId }
+            };
+
+            if (disableNotification)
+                parameters.Add("disable_notification", true);
+
+            return SendWebRequestAsync<bool>("pinChatMessage", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to unpin a message in a supergroup chat. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target supergroup</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns true on success</returns>
+        /// <see href="https://core.telegram.org/bots/api#unpinchatmessage"/>
+        public Task<bool> UnpinChatMessageAsync(ChatId chatId, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "chat_id", chatId }
+            };
+
+            return SendWebRequestAsync<bool>("unpinChatMessage", parameters, cancellationToken);
+        }
+        #endregion
+
+        #region Stickers
+        /// <summary>
+        /// Use this method to get a sticker set.
+        /// </summary>
+        /// <param name="name">Short name of the sticker set that is used in t.me/addstickers/ URLs (e.g., animals)</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>On success, a StickerSet object is returned.</returns>
+        /// <see href="https://core.telegram.org/bots/api#getstickerset"/>
+        public Task<StickerSet> GetStickerSetAsync(string name, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "name", name }
+            };
+
+            return SendWebRequestAsync<StickerSet>("getStickers", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to upload a .png file with a sticker for later use in createNewStickerSet and addStickerToSet methods (can be used multiple times).
+        /// </summary>
+        /// <param name="userId">User indentifier of sticker file owner</param>
+        /// <param name="pngSticker">Png image with the sticker, must be up to 512 kilobytes in size, dimensions must not exceed 512px, and either width or height must be exactly 512px.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns the uploaded File on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#uploadstickerfile"/>
+        public Task<File> UploadStickerFileAsync(int userId, FileToSend pngSticker, 
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "user_id", userId },
+                { "png_sticker", pngSticker }
+            };
+
+            return SendWebRequestAsync<File>("uploadStickerFile", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to create new sticker set owned by a user. The bot will be able to edit the created sticker set. 
+        /// </summary>
+        /// <param name="userId">User identifier of created sticker set owner</param>
+        /// <param name="name">Short name of sticker set, to be used in t.me/addstickers/ URLs (e.g., animals). Can contain only english letters, digits and underscores. Must begin with a letter, can't contain consecutive underscores and must end in “_by_&lt;bot_username&gt;”. &lt;bot_username&gt; is case insensitive. 1-64 characters.</param>
+        /// <param name="title">Sticker set title, 1-64 characters</param>
+        /// <param name="pngSticker">Png image with the sticker, must be up to 512 kilobytes in size, dimensions must not exceed 512px, and either width or height must be exactly 512px.</param>
+        /// <param name="emojis">One or more emoji corresponding to the sticker</param>
+        /// <param name="isMasks">Pass True, if a set of mask stickers should be created</param>
+        /// <param name="maskPosition">Position where the mask should be placed on faces</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns True on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#createnewstickerset"/>
+        public Task<bool> CreateNewStickerSetAsnyc(int userId, string name, string title, FileToSend pngSticker,
+            string emojis, bool isMasks = false, MaskPosition maskPosition = null, 
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "user_id", userId },
+                { "name", name },
+                { "title", title },
+                { "png_sticker", pngSticker },
+                { "emojis", emojis }
+            };
+
+            if (isMasks)
+                parameters.Add("is_masks", true);
+
+            if (maskPosition != null)
+                parameters.Add("mask_position", maskPosition);
+
+            return SendWebRequestAsync<bool>("createNewStickerSet", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to add a new sticker to a set created by the bot.
+        /// </summary>
+        /// <param name="userId">User identifier of sticker set owner</param>
+        /// <param name="name">Sticker set name</param>
+        /// <param name="pngSticker">Png image with the sticker, must be up to 512 kilobytes in size, dimensions must not exceed 512px, and either width or height must be exactly 512px.</param>
+        /// <param name="emojis">One or more emoji corresponding to the sticker</param>
+        /// <param name="maskPosition">Position where the mask should be placed on faces</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>True on success</returns>
+        /// <see href="https://core.telegram.org/bots/api#addstickertoset"/>
+        public Task<bool> AddStickerToSetAsync(int userId, string name, FileToSend pngSticker, string emojis, MaskPosition maskPosition = null,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "user_id", userId },
+                { "name", name },
+                { "png_sticker", pngSticker },
+                { "emojis", emojis }
+            };
+
+            if (maskPosition != null)
+                parameters.Add("mask_position", maskPosition);
+
+            return SendWebRequestAsync<bool>("addStickerToSet", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to move a sticker in a set created by the bot to a specific position.
+        /// </summary>
+        /// <param name="sticker">File identifier of the sticker</param>
+        /// <param name="position">New sticker position in the set, zero-based</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>True on success</returns>
+        /// <see href="https://core.telegram.org/bots/api#setstickerpositioninset"/>
+        public Task<bool> SetStickerPositionInSetAsync(string sticker, int position, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "sticker", sticker },
+                { "position", position }
+            };
+
+            return SendWebRequestAsync<bool>("setStickerPositionInSet", parameters, cancellationToken);
+        }
+
+        /// <summary>
+        /// Use this method to delete a sticker from a set created by the bot.
+        /// </summary>
+        /// <param name="sticker">File identifier of the sticker</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns True on success.</returns>
+        /// <see href="https://core.telegram.org/bots/api#deletestickerfromset"/>
+        public Task<bool> DeleteStickerFromSetAsync(string sticker, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            var parameters = new Dictionary<string, object>()
+            {
+                { "sticker", sticker }
+            };
+
+            return SendWebRequestAsync<bool>("deleteStickerFromSet", parameters, cancellationToken);
+        }
+        #endregion
+
         #region Support Methods - Private
 
         /// <summary>
@@ -1527,7 +1931,7 @@ namespace Telegram.Bot
         /// <param name="disableNotification">Sends the message silently. iOS users will not receive a notification, Android users will receive a notification with no sound.</param>
         /// <param name="replyToMessageId">Optional. If the message is a reply, ID of the original message</param>
         /// <param name="replyMarkup">Optional. Additional interface options. A JSON-serialized object for a custom reply keyboard, instructions to hide keyboard or to force a reply from the user.</param>
-        /// <param name="additionalParameters">Optional. if additional Parameters could bei send i.e. "disable_web_page_preview" in for a TextMessage</param>
+        /// <param name="additionalParameters">Optional. if additional Parameters could be send i.e. "disable_web_page_preview" in for a TextMessage</param>
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <returns>On success, the sent Message is returned.</returns>
         private Task<Message> SendMessageAsync(MessageType type, ChatId chatId, object content,
@@ -1607,7 +2011,7 @@ namespace Telegram.Bot
                 {
                     // Request with JSON data
 
-                    var payload = JsonConvert.SerializeObject(parameters);
+                    var payload = JsonConvert.SerializeObject(parameters, SerializerSettings);
 
                     var httpContent = new StringContent(payload, Encoding.UTF8, "application/json");
 
@@ -1618,7 +2022,7 @@ namespace Telegram.Bot
                 var responseString = await response.Content.ReadAsStringAsync()
                                                     .ConfigureAwait(false);
 
-                responseObject = JsonConvert.DeserializeObject<ApiResponse<T>>(responseString);
+                responseObject = JsonConvert.DeserializeObject<ApiResponse<T>>(responseString, SerializerSettings);
 
                 response.EnsureSuccessStatusCode();
             }
@@ -1635,10 +2039,11 @@ namespace Telegram.Bot
                 throw new ApiRequestException("Request timed out", 408, e);
             }
             catch (HttpRequestException e)
-                when (e.Message.Contains("400") || e.Message.Contains("403") || e.Message.Contains("409")) {}
+                when (e.Message.Contains("400") || e.Message.Contains("403") || e.Message.Contains("409"))
+            { }
 
             if (responseObject == null)
-                responseObject = new ApiResponse<T> {Ok = false, Message = "No response received"};
+                responseObject = new ApiResponse<T> { Ok = false, Message = "No response received" };
 
             if (!responseObject.Ok)
                 throw ApiRequestException.FromApiResponse(responseObject);
@@ -1652,15 +2057,10 @@ namespace Telegram.Bot
 
             switch (typeName)
             {
-                case "String":
-                case "Int32":
-                    return new StringContent(value.ToString(), Encoding.UTF8);
-                case "Boolean":
-                    return new StringContent((bool)value ? "true" : "false");
                 case "FileToSend":
                     return new StreamContent(((FileToSend)value).Content);
                 default:
-                    return new StringContent(JsonConvert.SerializeObject(value));
+                    return new StringContent(JsonConvert.SerializeObject(value, SerializerSettings).Trim('"'));
             }
         }
         #endregion