+++ src/Telegram.Bot/Types/InlineQueryResults/InlineQueryResult.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using Newtonsoft.Json;
-using Telegram.Bot.Converters;
 using Telegram.Bot.Types.InputMessageContents;
 using Telegram.Bot.Types.ReplyMarkups;