+++ src/Telegram.Bot/Types/Sticker.cs
@@ -4,6 +4,7 @@ namespace Telegram.Bot.Types
 {
     /// <summary>
     /// This object represents a sticker.
+    /// <see href="https://core.telegram.org/bots/api#sticker"/>
     /// </summary>
     [JsonObject(MemberSerialization.OptIn)]
     public class Sticker : File
@@ -31,5 +32,17 @@ namespace Telegram.Bot.Types
         /// </summary>
         [JsonProperty(PropertyName = "emoji", Required = Required.Default)]
         public string Emoji { get; set; }
+
+        /// <summary>
+        /// Optional. Name of the sticker set to which the sticker belongs
+        /// </summary>
+        [JsonProperty(PropertyName = "set_name", Required = Required.Default)]
+        public string SetName { get; set; }
+
+        /// <summary>
+        /// Optional. For mask stickers, the position where the mask should be placed
+        /// </summary>
+        [JsonProperty(PropertyName = "mask_position", Required = Required.Default)]
+        public MaskPosition MaskPosition { get; set; }
     }
 }